ref: rename utils to .ts
